requirements.txt
numpy
keras
pygame
sklearn
Pillow
sys
time